Resolved bug item #3508 (AVT2EXT GForge). Added code to check if loaded DICOM SEG object is null. If it is XIP Host throws exception which is displayed to the end used in the progress bar. Host is reset and ready to be used for a new query without the need to restart it.
The identified, most probable cause of it was that one of DICOM SEG objects (could be more than one) that were to be retrieved was deleted from AD_DICOM folder outside of the XIP Host or AD interface. It means database was corrupted and referenced from the database AD_DICOM directory did not contain specified DICOM SEG object. 

SVN-Revision: 416